Fix template paths
