+++ backend/openai_tools.py
@@ -1,0 +1,114 @@
+import json
+import pathlib
+import subprocess
+
+from openai import OpenAI
+
+from .utils import within_root, run_cmd
+
+
+def get_client(provider: str):
+    if provider.lower() == "ollama":
+        return OpenAI(base_url="http://localhost:11434/v1", api_key="ollama")
+    return OpenAI()
+
+# ---------- helpers ---------- #
+
+TOOLS = [
+    {
+        "type": "function",
+        "function": {
+            "name": "write_file",
+            "description": "Create or overwrite a file",
+            "parameters": {
+                "type": "object",
+                "properties": {
+                    "filename": {"type": "string"},
+                    "content": {"type": "string"},
+                },
+                "required": ["filename", "content"],
+            },
+        },
+    },
+    {
+        "type": "function",
+        "function": {
+            "name": "read_file",
+            "description": "Return the contents of a text file",
+            "parameters": {
+                "type": "object",
+                "properties": {"filename": {"type": "string"}},
+                "required": ["filename"],
+            },
+        },
+    },
+    {
+        "type": "function",
+        "function": {
+            "name": "write_command",
+            "description": "Execute a Unix command and capture stdout/stderr",
+            "parameters": {
+                "type": "object",
+                "properties": {"command": {"type": "string"}},
+                "required": ["command"],
+            },
+        },
+    },
+    {
+        "type": "function",
+        "function": {
+            "name": "change_file",
+            "description": (
+                "Apply a git patch to a file. Call read_file first to get"
+                " the current content and then send a unified diff patch."
+            ),
+            "parameters": {
+                "type": "object",
+                "properties": {
+                    "filename": {"type": "string"},
+                    "patch": {"type": "string"},
+                },
+                "required": ["filename", "patch"],
+            },
+        },
+    },
+]
+
+
+def write_file(filename, content):
+    path = pathlib.Path(filename).expanduser()
+    path.parent.mkdir(parents=True, exist_ok=True)
+    path.write_text(content)
+    return f"Wrote {path} ({len(content)} bytes)."
+
+
+def read_file(filename):
+    p = pathlib.Path(filename).expanduser()
+    return p.read_text() if p.exists() else f"{p} not found."
+
+
+def change_file(filename: str, patch: str):
+    path = pathlib.Path(filename).expanduser()
+    if not within_root(path):
+        return "Blocked: path outside working directory."
+    try:
+        res = subprocess.run(
+            ["git", "apply", "-"],
+            input=patch,
+            text=True,
+            capture_output=True,
+        )
+        if res.returncode != 0:
+            content = path.read_text() if path.exists() else ""
+            return f"Patch failed:\n{res.stderr}\nCurrent file:\n{content}"
+        return f"Patch applied to {path}."
+    except Exception as exc:
+        return f"Error applying patch: {exc}"
+
+# map tool names to callables
+TOOL_FUNCS = {
+    "write_file": write_file,
+    "read_file": read_file,
+    "write_command": run_cmd,
+    "change_file": change_file,
+}
